% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom_05_p_val}
\alias{binom_05_p_val}
\title{vectorized binomial p value}
\usage{
binom_05_p_val(a, b)
}
\arguments{
\item{a}{number of successes}

\item{b}{number of failures (defaults to n - a)}
}
\value{
p value of binomial test
}
\description{
vectorized binomial p value
}
