% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom_p_val}
\alias{binom_p_val}
\title{vectorized binomial p value}
\usage{
binom_p_val(a, b = n - a, n = a + b, null_prob = 0.5)
}
\arguments{
\item{a}{number of successes}

\item{b}{number of failures (defaults to n - a)}

\item{n}{number of trials (defaults to a + b)}

\item{null_prob}{prob of success under the null hypothesis, default 0.5}
}
\value{
p value of binomial test
}
\description{
vectorized binomial p value
}
