% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_controls.R
\name{check_matching_feasibility}
\alias{check_matching_feasibility}
\title{Check if matching is feasible and return details of any problems}
\usage{
check_matching_feasibility(demand_df, supply_df, match_vars)
}
\description{
Check if matching is feasible and return details of any problems
}
