% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{checkpoint}
\alias{checkpoint}
\title{checkpoint saves/loads data from local disk or google bucket}
\usage{
checkpoint(x, dir, local_disk = TRUE, google_bucket = TRUE)
}
\arguments{
\item{x}{the object in R}

\item{dir}{the save directory}

\item{local_disk}{binary indicating whether to save to local disk}

\item{google_bucket}{binary indicating whether to save to google bucket}
}
\value{
TRUE
}
\description{
checkpoint saves/loads data from local disk or google bucket
}
