% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonest_values.R
\name{commonest_values}
\alias{commonest_values}
\title{commonest values}
\usage{
commonest_values(v, n = 5)
}
\arguments{
\item{v}{the vector}

\item{n}{how many values to return}
}
\value{
the n most common values in v
}
\description{
commonest values
}
