% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CortyKit2-package.R
\docType{package}
\name{CortyKit2-package}
\alias{CortyKit2}
\alias{CortyKit2-package}
\title{CortyKit2: Robert Corty's Utility Functions}
\description{
This package offers a variety of semi-related utility functions that I've had to write as part of my work. It's package to facilitate their use across computing environments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bicklab/CortyKit2}
  \item Report bugs at \url{https://github.com/bicklab/CortyKit2/issues}
}

}
\author{
\strong{Maintainer}: Robert Corty \email{rcorty@gmail.com} (\href{https://orcid.org/0000-0002-6787-9051}{ORCID})

}
\keyword{internal}
