% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatic_mutation_processing.R
\name{filter_sm_one_gene}
\alias{filter_sm_one_gene}
\title{Filter SM one gene}
\usage{
filter_sm_one_gene(sm_table, demog)
}
\arguments{
\item{sm_table}{somatic mutation tibble}

\item{demog}{demographics tibble}
}
\value{
filtered SM table
}
\description{
Filter SM one gene
}
