% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_controls.R
\name{get_control_demand}
\alias{get_control_demand}
\title{Calculate required controls for each matching group}
\usage{
get_control_demand(data, case_ids, match_vars, ccr)
}
\description{
Calculate required controls for each matching group
}
