% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_controls.R
\name{get_control_ids}
\alias{get_control_ids}
\title{get control ids}
\usage{
get_control_ids(
  case_ids,
  covariates,
  exclude_ids = NULL,
  control_case_ratio = 10,
  verbose = FALSE
)
}
\arguments{
\item{case_ids}{the case ids}

\item{covariates}{covariates, some of which will be matched on}

\item{exclude_ids}{ids that user does not want to consider as possible controls}

\item{control_case_ratio}{number of controls per case, default to 10}

\item{verbose}{whether to output what ends up getting matched on}
}
\value{
a vector of control ids
}
\description{
get control ids
}
