% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_controls.R
\name{get_control_supply}
\alias{get_control_supply}
\title{Calculate available controls for each matching group}
\usage{
get_control_supply(data, case_ids, match_vars)
}
\description{
Calculate available controls for each matching group
}
