% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrevs.R
\name{\%nin\%}
\alias{\%nin\%}
\title{nin: not in}
\usage{
a \%nin\% b
}
\arguments{
\item{a}{one vector}

\item{b}{another vector}
}
\value{
a binary vector of whether each entry in A is not in B (the converse of \%in\%)
}
\description{
nin: not in
}
