% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsutil.R
\name{gscp}
\alias{gscp}
\title{gscp}
\usage{
gscp(src, dest, echo = TRUE)
}
\arguments{
\item{src}{source}

\item{dest}{destination}

\item{echo}{whether to print the command to screen before submitting to terminal#'}
}
\value{
terminal output from gsutil command
}
\description{
gscp
}
