% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsutil.R
\name{gsls}
\alias{gsls}
\title{gsls}
\usage{
gsls(uri, echo = TRUE)
}
\arguments{
\item{uri}{the URI to list}

\item{echo}{whether to print the command to screen before submitting to terminal}
}
\value{
terminal output from gsutil command
}
\description{
gsls
}
