% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsutil.R
\name{gsutil}
\alias{gsutil}
\title{gsutil}
\usage{
gsutil(text, echo = FALSE)
}
\arguments{
\item{text}{the gsutil commands}

\item{echo}{whether to print the command to screen before submitting to terminal}
}
\value{
terminal output from gsutil command
}
\description{
gsutil
}
