% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatic_mutation_plotting.R
\name{make_age_VAF_plot}
\alias{make_age_VAF_plot}
\title{make age-vaf plot}
\usage{
make_age_VAF_plot(cases)
}
\arguments{
\item{cases}{tibble of cases}
}
\value{
the plot
}
\description{
make age-vaf plot
}
