% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatic_mutation_plotting.R
\name{make_prevalence_plot}
\alias{make_prevalence_plot}
\title{make prevalence plot}
\usage{
make_prevalence_plot(cases, demog)
}
\arguments{
\item{cases}{tibble of cases}

\item{demog}{tibble of demographic info}
}
\value{
the plot
}
\description{
make prevalence plot
}
