% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_controls.R
\name{match_cases_to_controls}
\alias{match_cases_to_controls}
\title{Match cases to controls from a dataset based on matching variables.}
\usage{
match_cases_to_controls(
  case_ids,
  match_variables,
  data,
  sampling_seed = 37232,
  control_case_ratio = 5
)
}
\arguments{
\item{case_ids}{the IDs of the cases}

\item{match_variables}{the variables to match on}

\item{data}{The data about cases and potential controls.
It must contain match_variables as columns and have a column called
'person_id' that contains all values of case_ids}

\item{sampling_seed}{seed used for random sampling of controls from data
(default is ZIP code of VUMC)}

\item{control_case_ratio}{how many controls do we want per case?}
}
\value{
A tibble with columns \code{case_id} and \code{control_id}, one row per
matched case-control pair. Each control appears at most once.
}
\description{
Match cases to controls from a dataset based on matching variables.
}
