% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_case_control.R
\name{phewas_case_control}
\alias{phewas_case_control}
\title{phewas for case-control design}
\usage{
phewas_case_control(
  case_ids,
  control_ids,
  phecode_counts,
  shrinkage_pseudocounts = 5
)
}
\arguments{
\item{case_ids}{patient ID's of cases}

\item{control_ids}{patient ID's of controls}

\item{phecode_counts}{tibble of phecodes}

\item{shrinkage_pseudocounts}{how many pseudocounts to add to true counts of
cases/controls with/without disease, default to 5 to decrease sensitivity
to rare occurrences}
}
\value{
odds ratios and p-values for each phecode
}
\description{
phewas for case-control design
}
