% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqunif}
\alias{qqunif}
\title{qqunif}
\usage{
qqunif(ps)
}
\arguments{
\item{ps}{the p-values (between 0 and 1)}
}
\value{
a QQ plot
}
\description{
simple qq plotting function that assumes theoretical distribution is uniform
}
