% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonest_values.R
\name{second_highest_val}
\alias{second_highest_val}
\title{second highest value of vector}
\usage{
second_highest_val(x)
}
\arguments{
\item{x}{the vector}
}
\value{
the second highest value in x
}
\description{
second highest value of vector
}
