% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_phewas.R
\name{surv_phewas}
\alias{surv_phewas}
\title{survival phewas}
\usage{
surv_phewas(covars, phecodes)
}
\arguments{
\item{covars}{the covariates to use with each phecode}

\item{phecodes}{the phecodes to scan over}
}
\value{
a table of effects and p-values for each phecode
}
\description{
survival phewas
}
