% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrevs.R
\name{years_between}
\alias{years_between}
\title{years_between}
\usage{
years_between(start, end)
}
\arguments{
\item{start}{start date}

\item{end}{end date}
}
\value{
the number of years between the two dates
}
\description{
years_between
}
