% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_controls.R
\name{get_control_ids}
\alias{get_control_ids}
\title{Main function to get control IDs}
\usage{
get_control_ids(
  case_ids,
  match_variables,
  data,
  exclude_ids = NULL,
  control_case_ratio = 5
)
}
\description{
Main function to get control IDs
}
